Add HTTP assertions
Added support to assert on a HTTP response.
The ResponseImpl verifies on a HTTP response object, which is
the returned object from the ClientImpl's send method.

It's also possible to assert on only specific parts of the
HTTP response,, for example the request. This can be done
in two different ways:
By using the ResponseImpl and set the assertion for the request, by
calling responseNodeImpl.setRequest(..).
By creating a RequestNodeImpl and verify on httpResponse.getRequest.

Since this is a big commit, tests will be added in a follow up
commit, together with possible fixes.