Add server support
Added support for a simple HTTP server.
The server can be created through the factory class.

Other new public interfaces are Context, Request, Response, Streams
and some builders for these.
These are all related to the server where the context is configuration
to be able to setup a a handler for incoming requests.
The request is the incoming request, the response is the response the
server should respond on the incoming request.

The streams is a wrapper around an output and input stream.

Since this commis is huge, tests will be added in an upcoming commit.