Add tests in server package
Added all missing tests for the server package.

Also fixed a bug where an never-ending loop occured.
The reason was that the boolean, checking if the thread
should wait or not, was set to false, which means the
thread never waited. This resulted in the loop that could
never end.