Add IT to Simple server
Added IT to the Simple class in the server package.
Also changed the Factory to return the Simple interface,
rather than the Server interface.

This is needed since it should not be mandatory to use
a fixture to use a server. This could of course be done with a cast,
but that is not very clear.

Also changed SimpleTest, since it's no longer needed to do the cast.